# proto/a13n-service/openapi.json
{
  "components": {
    "schemas": {
...
      "Authentication": {
        "additionalProperties": false,
        "properties": {
          "cases": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AuthenticationCase"
            },
            "title": "Cases",
            "type": "array"
          },
          "mode": {
            "$ref": "#/components/schemas/CredentialMode",
            "default": "required"
          }
        },
        "title": "Authentication",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JsonValue": {},
...
      "ModelConfig": {
        "additionalProperties": false,
        "properties": {
          "context_window": {
            "maximum": 10000000.0,
            "minimum": 1.0,
            "title": "Context Window",
            "type": "integer"
          },
          "max_tokens": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "model_api": {
            "enum": [
              "openai.responses",
              "openai.chat_completions"
            ],
            "title": "Model Api",
            "type": "string"
          },
          "model_name": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "top_p": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top P"
          }
        },
        "required": [
          "model_name",
          "model_api",
          "context_window"
        ],
        "title": "ModelConfig",
        "type": "object"
      },
      "ModelCreate": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/ModelConfig"
          },
          "key": {
            "pattern": "^[a-z0-9][a-z0-9_-]{0,127}$",
            "title": "Key",
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pricing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelPricingEntry-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider_id": {
            "maxLength": 72,
            "pattern": "^[a-z][a-z0-9]{1,7}_[a-z0-9]{16,64}$",
            "title": "Provider Id",
            "type": "string"
          },
          "workspace_id": {
            "anyOf": [
              {
                "maxLength": 72,
                "pattern": "^[a-z][a-z0-9]{1,7}_[a-z0-9]{16,64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "provider_id",
          "key",
          "name",
          "config"
        ],
        "title": "ModelCreate",
        "type": "object"
      },
      "ModelPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModelView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ModelPage",
        "type": "object"
      },
      "ModelPriceRule-Input": {
        "additionalProperties": false,
        "description": "One complete set of prices and the condition selecting it.",
        "properties": {
          "constraint": {
            "$ref": "#/components/schemas/PricingConstraint"
          },
          "prices": {
            "items": {
              "$ref": "#/components/schemas/PriceComponent-Input"
            },
            "maxItems": 128,
            "minItems": 1,
            "title": "Prices",
            "type": "array"
          },
          "rule_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Rule Id",
            "type": "string"
          }
        },
        "required": [
          "rule_id",
          "prices"
        ],
        "title": "ModelPriceRule",
        "type": "object"
      },
      "ModelPriceRule-Output": {
        "additionalProperties": false,
        "description": "One complete set of prices and the condition selecting it.",
        "properties": {
          "constraint": {
            "$ref": "#/components/schemas/PricingConstraint"
          },
          "prices": {
            "items": {
              "$ref": "#/components/schemas/PriceComponent-Output"
            },
            "maxItems": 128,
            "minItems": 1,
            "title": "Prices",
            "type": "array"
          },
          "rule_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Rule Id",
            "type": "string"
          }
        },
        "required": [
          "rule_id",
          "prices"
        ],
        "title": "ModelPriceRule",
        "type": "object"
      },
      "ModelPricingEntry-Input": {
        "additionalProperties": false,
        "description": "Complete pricing declaration for one provider-qualified model.",
        "properties": {
          "context_window": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Window"
          },
          "model": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/ModelPriceRule-Input"
            },
            "maxItems": 128,
            "minItems": 1,
            "title": "Rules",
            "type": "array"
          },
          "source": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Source",
            "type": "string"
          },
          "source_revision": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Source Revision",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "required": [
          "provider",
          "model",
          "rules",
          "source",
          "source_revision"
        ],
        "title": "ModelPricingEntry",
        "type": "object"
      },
      "ModelPricingEntry-Output": {
        "additionalProperties": false,
        "description": "Complete pricing declaration for one provider-qualified model.",
        "properties": {
          "context_window": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Window"
          },
          "model": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/ModelPriceRule-Output"
            },
            "maxItems": 128,
            "minItems": 1,
            "title": "Rules",
            "type": "array"
          },
          "source": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Source",
            "type": "string"
          },
          "source_revision": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Source Revision",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "required": [
          "provider",
          "model",
          "rules",
          "source",
          "source_revision"
        ],
        "title": "ModelPricingEntry",
        "type": "object"
      },
      "ModelView": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/ModelConfig"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "pricing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelPricingEntry-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "id",
          "organization_id",
          "workspace_id",
          "provider_id",
          "key",
          "name",
          "config",
          "pricing",
          "enabled",
          "version"
        ],
        "title": "ModelView",
        "type": "object"
      },
...
      "PriceComponent-Output": {
        "additionalProperties": false,
        "description": "One genai-prices usage dimension and its USD unit price.",
        "properties": {
          "price": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price",
            "type": "string"
          },
          "price_key": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Price Key",
            "type": "string"
          },
          "tiers": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PriceTier-Output"
            },
            "maxItems": 32,
            "title": "Tiers",
            "type": "array"
          }
        },
        "required": [
          "price_key",
          "price"
        ],
        "title": "PriceComponent",
        "type": "object"
      },
...
      "ProviderCreate": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Config",
            "type": "object"
          },
          "credential": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "type": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Type",
            "type": "string"
          },
          "workspace_id": {
            "anyOf": [
              {
                "maxLength": 72,
                "pattern": "^[a-z][a-z0-9]{1,7}_[a-z0-9]{16,64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "type",
          "name",
          "config"
        ],
        "title": "ProviderCreate",
        "type": "object"
      },
      "ProviderPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProviderView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ProviderPage",
        "type": "object"
      },
      "ProviderType": {
        "properties": {
          "authentication": {
            "$ref": "#/components/schemas/Authentication"
          },
          "configuration_schema": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Configuration Schema",
            "type": "object"
          },
          "credential_schema": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Schema"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "setup_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setup Url"
          },
          "supported_model_apis": {
            "items": {
              "type": "string"
            },
            "title": "Supported Model Apis",
            "type": "array"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "display_name",
          "configuration_schema",
          "credential_schema",
          "authentication",
          "supported_model_apis",
          "setup_url"
        ],
        "title": "ProviderType",
        "type": "object"
      },
      "ProviderTypePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProviderType"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ProviderTypePage",
        "type": "object"
      },
      "ProviderView": {
        "properties": {
          "config": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Config",
            "type": "object"
          },
          "credential_configured": {
            "title": "Credential Configured",
            "type": "boolean"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "id",
          "organization_id",
          "workspace_id",
          "type",
          "name",
          "config",
          "credential_configured",
          "enabled",
          "version"
        ],
        "title": "ProviderView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "a13n Service",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/organizations/{organization_id}/model-providers": {
      "get": {
        "operationId": "list_providers_api_v1_organizations__organization_id__model_providers_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Providers",
        "tags": [
          "models"
        ]
      },
      "post": {
        "operationId": "create_provider_api_v1_organizations__organization_id__model_providers_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Provider",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/model-providers/{provider_id}": {
      "get": {
        "operationId": "get_provider_api_v1_organizations__organization_id__model_providers__provider_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Provider",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/models": {
      "get": {
        "operationId": "list_models_api_v1_organizations__organization_id__models_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Models",
        "tags": [
          "models"
        ]
      },
      "post": {
        "operationId": "create_model_api_v1_organizations__organization_id__models_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Model",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/models/{model_id}": {
      "get": {
        "operationId": "get_model_api_v1_organizations__organization_id__models__model_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/provider-types/model": {
      "get": {
        "operationId": "provider_types_api_v1_provider_types_model_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderTypePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provider Types",
        "tags": [
          "models"
        ]
      }
    },
...
  }
}